+++ src/main/java/classloader/constantpool/info/InterfaceMethodrefInfo.java
@@ -7,10 +7,10 @@ package classloader.constantpool.info;
  * Created on 2020-02-15
  */
 public class InterfaceMethodrefInfo extends ConstantPoolInfo {
-    private short classIndex;
-    private short nameAndTypeIndex;
+    private int classIndex;
+    private int nameAndTypeIndex;
 
-    public InterfaceMethodrefInfo(short classIndex, short nameAndTypeIndex) {
+    public InterfaceMethodrefInfo(int classIndex, int nameAndTypeIndex) {
         this.classIndex = classIndex;
         this.nameAndTypeIndex = nameAndTypeIndex;
         super.tag = ConstantPoolInfo.INTERFACE_METHOD_REF;
