+++ src/main/java/classloader/constantpool/info/MethodHandleInfo.java
@@ -7,10 +7,10 @@ package classloader.constantpool.info;
  * Created on 2020-02-15
  */
 public class MethodHandleInfo extends ConstantPoolInfo{
-    private byte referenceKind;
-    private short referenceIndex;
+    private int referenceKind;
+    private int referenceIndex;
 
-    public MethodHandleInfo(byte referenceKind, short referenceIndex) {
+    public MethodHandleInfo(int referenceKind, int referenceIndex) {
         this.referenceKind = referenceKind;
         this.referenceIndex = referenceIndex;
         super.tag = ConstantPoolInfo.METHOD_HANDLE;
