+++ src/main/java/classloader/constantpool/info/MethodTypeInfo.java
@@ -7,9 +7,9 @@ package classloader.constantpool.info;
  * Created on 2020-02-15
  */
 public class MethodTypeInfo extends ConstantPoolInfo {
-    private short descriptorIndex;
+    private int descriptorIndex;
 
-    public MethodTypeInfo(short descriptorIndex) {
+    public MethodTypeInfo(int descriptorIndex) {
         this.descriptorIndex = descriptorIndex;
         super.tag = ConstantPoolInfo.METHOD_TYPE;
     }
