+++ src/main/java/classloader/constantpool/info/NameAndTypeInfo.java
@@ -9,10 +9,10 @@ import com.sun.tools.classfile.ConstantPool;
  * Created on 2020-02-15
  */
 public class NameAndTypeInfo extends ConstantPoolInfo {
-    private short nameIndex;
-    private short descriptorIndex;
+    private int nameIndex;
+    private int descriptorIndex;
 
-    public NameAndTypeInfo(short nameIndex, short descriptorIndex) {
+    public NameAndTypeInfo(int nameIndex, int descriptorIndex) {
         this.nameIndex = nameIndex;
         this.descriptorIndex = descriptorIndex;
         super.tag = ConstantPoolInfo.NAME_AND_TYPE;
